requirements.txt
fastapi~=0.109.2
openai~=1.109.1
langchain-core==1.2.7
langchain==1.2.6
langchain-deepseek
langchain-mcp-adapters==0.2.1
langgraph==1.0.6
pydantic~=2.12.2
sqlalchemy~=2.0.44
uvicorn~=0.37.0
dashscope~=1.24.8
python-dotenv>=1.1.0
fastmcp==2.2.1
langchain-text-splitters
mcp